+++ packages/web/src/lib/stores/playlist.ts
@@ -1,4 +1,3 @@
-import { createMemo } from 'solid-js';
 import { createSyncedPersistedSignal } from './base';
 
 export type PlaylistItem = {