+++ packages/web/src/views/BroadcastScreen.tsx
@@ -1,10 +1,10 @@
 import styles from './BroadcastScreen.module.scss';
 import { createEffect, createMemo, createSignal, Match, onCleanup, onMount, Show, Switch } from 'solid-js';
-import Ticker from '../components/Ticker';
-import Banner from '../components/Banner';
-import { ErrorDisplay } from '../components/ErrorDisplay';
+import Ticker from '../components/Ticker.tsx';
+import Banner from '../components/Banner.tsx';
+import { ErrorDisplay } from '../components/ErrorDisplay.tsx';
 import { getFileAndType } from '../lib/stores/file-store.ts';
-import { endCurrentCall, setupQRCodeFlow } from '../lib/qr2phone2stream';
+import { endCurrentCall, setupQRCodeFlow } from '../lib/qr2phone2stream.ts';
 import { MediaChangeParams, registerOnEndCallHandler, onMediaChange } from '../lib/inter-tab-comms.ts';
 import {
     mediaStream,